requirements.txt
flask
flask-bootstrap
newsapi-python
transformers